dga-reports/requirements.txt
google-cloud-tasks
google-cloud-bigquery
sqlalchemy~=1.4.31
pg8000
pymongo
requests